Update index based on client's request
It should be ‘Citra’, remove the word ‘Lab’
Take off the second email (youthandinnovation@gmail.com). Leave only innovations.undplk@gmail.com
In the ‘Our Partners’ section, take off LirneAsia and Pulse Lab. Will add more later. For now it should only be COSTI and the Ministry
In the footer (for reference, attachment titled ‘Footer’) should not include the social media handles as just above that it gives the handles